Adds a mailbox implementation
A mailbox is a spsc channel with a capacity of one.
The send side will not block when the capacity is reached but instead returns an error.
The send side will spin when trying to acquire the mutex instead of waiting a condvar.
The receive side will block when the mailbox is empty and wait on a condvar for a notification